+++ contracts-c/toolbox/filter_on_xrp.c
@@ -7,11 +7,11 @@
 int64_t hook(uint32_t reserved) {
     TRACESTR("filter_on_xrp: Start.");
 
-    // FILTER ON AMOUNT: XRP
+    // FILTER ON AMOUNT: XAH
     uint8_t amount_buffer[48];
     int64_t amount_len = otxn_field(SBUF(amount_buffer), sfAmount);
     if (amount_len != 8) {
-        rollback(SBUF("filter_on_xrp: Ignoring non XRP Transaction"), __LINE__);
+        rollback(SBUF("filter_on_xrp: Ignoring non XAH Transaction"), __LINE__);
     }
 
     int64_t amount_drops = AMOUNT_TO_DROPS(amount_buffer);