+++ test/integration-c/hooks/toolbox/filterOnXrp.test.ts
@@ -98,7 +98,7 @@ describe('filterOnXrp', () => {
     } catch (error: unknown) {
       if (error instanceof Error) {
         expect(error.message).toEqual(
-          'filter_on_xrp: Ignoring non XRP Transaction'
+          'filter_on_xrp: Ignoring non XAH Transaction'
         )
       }
     }